fix: vue form with choice expanded (#1802)
| Q             | A
| ------------- | ---
| Bug fix?      | yes
| New feature?  | no
| License       | MIT

Choice children, when choice has the option expanded, have different full names as other forms. In php, the `ChoiceType` will change the full name in `finishView`. So on the js site, we can't use the `updateFullName` in `ListForm`, because we assume that every full name of a form has the same build rules. So we have to delegate the rebuilding of the full name to the model, so it can overwrite its behaviour if needed.