[MultiTenancyBundle] Fix call on null error exception if Resolver is used from Command (#1844)
| Q            | A
|--------------| ---
| Bug fix?     | yes
| New feature? | no
| BC-Break?    | no
| Backport     | 0.10
| Tickets      | Fix #... <!-- prefix each issue number with "Fix #", if any -->
| License      | MIT

If e.g. TenantManager->getTenant was called from a CLI Command during resolver->getTenant no Request is provided in SessionResolver.
Instead of throwing an Error due to a call on a null object, return null and see if any of the remaining resolvers can resolve the tenant and finally return null if none was found.